moved ua-msi parsing as per review
